+++ extensions/zen-browser/src/components/ZenListEntries.tsx
@@ -1,11 +1,12 @@
 import { Icon, List } from "@raycast/api";
 import { ZenActions } from "./index";
-import { HistoryEntry, Tab } from "../interfaces";
+import { HistoryEntry, Tab, WorkspaceEntry } from "../interfaces";
 import { getFavicon } from "@raycast/utils";
 
 export class ZenListEntries {
   public static NewTabEntry = NewTabEntry;
   public static HistoryEntry = HistoryListEntry;
+  public static WorkspaceEntry = WorkspaceListEntry;
 }
 
 function NewTabEntry({ searchText }: { searchText?: string }) {
@@ -14,6 +15,25 @@ function NewTabEntry({ searchText }: { searchText?: string }) {
       title={!searchText ? "Open Empty Tab" : `Search "${searchText}"`}
       icon={{ source: !searchText ? Icon.Plus : Icon.MagnifyingGlass }}
       actions={<ZenActions.NewTab query={searchText} />}
+    />
+  );
+}
+
+function WorkspaceListEntry({ workspace }: { workspace: WorkspaceEntry }) {
+  const shortcutAccessory = workspace.shortcut
+    ? {
+        text: `${workspace.shortcut.modifiers
+          .map((m) => m[0].toUpperCase() + m.slice(1))
+          .join("+")}+${workspace.shortcut.key.toUpperCase()}`,
+      }
+    : { text: "Assign a Shortcut", tooltip: "Shortcuts are necessary to make this feature work" };
+
+  return (
+    <List.Item
+      title={workspace.name}
+      accessories={[{ text: workspace.is_default ? "Default" : "" }, shortcutAccessory]}
+      actions={<ZenActions.WorkspaceItem workspace={workspace} />}
+      icon={{ source: workspace.icon }}
     />
   );
 }