+++ extensions/zen-browser/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Zen Changelog
 
+## [Added Open Workspace Command] - 2025-09-14
+
+- Added a command to open a specific workspace in Zen Browser.
+- Added a hook to fetch user shortcuts.
+- Improved retry mechanism for SQL queries.
 ## [Windows Support] - 2025-09-16
 
 - Added Windows Support